refactor: changes schedulers to use floating point values instead of integers

+++ include/utils.hpp
@@ -3,6 +3,6 @@
 
 #include "types.hpp"
 
-int endTime(const ExecutionBlock &exec);
+float endTime(const ExecutionBlock &exec);
 
 std::vector<size_t> orderOfArrival(const std::vector<Process> &processes);
