+++ src/core/utils.cpp
@@ -5,7 +5,7 @@
 #include "types.hpp"
 #include "utils.hpp"
 
-int endTime(const ExecutionBlock &exec) {
+float endTime(const ExecutionBlock &exec) {
     return exec.startTime + exec.duration;
 }
